[VL] Rename OASPackageBridges
