+++ gluten-core/src/main/scala/org/apache/gluten/backendsapi/IteratorApi.scala
@@ -27,7 +27,7 @@ import org.apache.spark._
 import org.apache.spark.sql.connector.read.InputPartition
 import org.apache.spark.sql.execution.metric.SQLMetric
 import org.apache.spark.sql.types.StructType
-import org.apache.spark.sql.utils.OASPackageBridge.InputMetricsWrapper
+import org.apache.spark.sql.utils.SparkInputMetricsUtil.InputMetricsWrapper
 import org.apache.spark.sql.vectorized.ColumnarBatch
 
 trait IteratorApi {
